Moved WebSocketServer into Test.Common.
